+++ client/src-tauri/src/lib.rs
@@ -97,6 +97,7 @@ pub fn run() {
     tauri::Builder::default()
         .manage(PlaybackQueue::default())
         .manage(PlaybackSessionStore::default())
+        .plugin(tauri_plugin_clipboard_manager::init())
         .plugin(tauri_plugin_dialog::init())
         .plugin(tauri_plugin_opener::init())
         .plugin(tauri_plugin_os::init())